funk: add placeholder accounts_lru_prev_idx/accounts_lru_next_idx rec fields

+++ src/app/shared/fd_config.h
@@ -399,7 +399,7 @@ struct fd_config {
     struct {
       char  capture[ PATH_MAX ];
       char  funk_checkpt[ PATH_MAX ];
-      ulong funk_rec_max;
+      uint  funk_rec_max;
       ulong funk_sz_gb;
       ulong funk_txn_max;
       char  funk_file[ PATH_MAX ];
